메인게시판 전체적으로 수정 완료!
