+++ src/main/resources/static/js/detaile.js
@@ -2,5 +2,5 @@ function detailepath(path, name){
 
     $('#mainContent').children().remove();
     $('#mainContent').load('/'+path+'/'+name);
-    
+
 }
