+++ src/main/resources/static/js/modify.js
@@ -1,4 +1,5 @@
-function modify(path, name){
+
+function boardmodify(){
 
     const $mainboardBno = $('#mainboardBno');
     const $mainboardTitle = $('#mainboardTitle');
@@ -7,19 +8,21 @@ function modify(path, name){
 
     $.ajax({
         type: 'post',
-        url: document.modifyfrm.action,
+        url: '/main/modify',
         data: {
             mainboardBno: $mainboardBno.val(),
             mainboardTitle: $mainboardTitle.val(),
-            mainboardContent: $mainboardContent.val()
+            mainboardContent: $mainboardContent.val(),
+            mainboardWriter: $mainboardWriter.val(),
         },
         success: function(data){
             alert('글수정에 성공하였습니다!');
-            $('#mainContent').children().remove();
-            $('#mainContent').load('/'+path+'/'+name+'?mainboardWriter='+$mainboardWriter.val());
+            window.location.href='/main/mainboard';
         },
         error: function(error){
+
             alert('글수정에 실패하였습니다!');
+
         }
 
     })
