+++ src/main/webapp/WEB-INF/views/main/main.jsp
@@ -38,10 +38,7 @@
             </c:if>
             <br>
             <br>
-            
-            <form action="/logout" method="post">
-                <button class="btn btn-primary" style="width: 186px;" href="/logout">로그아웃</button>
-            </form>
+            <a class="btn btn-primary" href="http://localhost:8097/logout" style="width: 186px;">로그아웃</a>
           </a>
     
         </li>
@@ -166,6 +163,7 @@
           </nav>
           <!-- End of Topbar -->
           <div id="mainContent" class="container-fluid">
+              <h1>myERP</h1>
           </div>
 
         </div>
